+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Support for propagating errors via an `assign_errors` helper ([#10](https://github.com/svycal/inertia-phoenix/issues/10))
 - Preservation of assigned errors across redirects ([#10](https://github.com/svycal/inertia-phoenix/issues/10))
 - Setup external redirects properly for Inertia requests ([#11](https://github.com/svycal/inertia-phoenix/issues/11))
+- Forward flash contents across forced refreshes ([#13](https://github.com/svycal/inertia-phoenix/issues/13))
+- Automatically pass Phoenix flash data via the `flash` prop
 
 ## 0.4.0